Crowd_count/CrowdCounting-using-video-analysis--main/auth_utils.py
import jwt
from functools import wraps
from flask import request, redirect, current_app

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.cookies.get("token")

        if not token:
            return redirect("/")

        try:
            jwt.decode(
                token,
                current_app.config["SECRET_KEY"],
                algorithms=["HS256"]
            )
        except:
            return redirect("/")

        return f(*args, **kwargs)
    return decorated


def admin_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.cookies.get("token")
        data = jwt.decode(
            token,
            current_app.config["SECRET_KEY"],
            algorithms=["HS256"]
        )

        if data.get("role") != "admin":
            return "Access Denied", 403

        return f(*args, **kwargs)
    return decorated
